requirements.txt
llmauditor>=1.1.3
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.0.0
streamlit>=1.28.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
schedule>=1.2.0